requirements.txt

pytest
mock
luma.oled
netifaces
python-mpd2import
wifiinfo